move useMetabaseQuery tests into tests folder
